requirements.txt
# Core
Django==5.1.3
django-environ==0.11.2
gunicorn==23.0.0
psycopg2-binary==2.9.10

# Rest apis
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
drf-yasg==1.21.8
whitenoise==6.8.2

# Certificates
cryptography==43.0.3

# Network
portscan==1.1

# Ring Camera
ring_doorbell==0.9.13

# Docker
docker==7.1.0

# Celery
celery==5.4.0
django-celery-beat==2.8.1
# Redis
redis==5.2.1

# Selenium
selenium==4.27.1

# Onvif-zeep
onvif_zeep==0.2.12

pgvector==0.3.6

psutil==5.9.8
paho-mqtt==2.1.0

# wake word requirement
openai-whisper==20250625

torch==2.9.1
numpy==2.2.6
scipy==1.15.3
sounddevice==0.5.3
soundfile==0.13.1